Replaced List with CList
List suffers from many drawbacks:
* It can hold only pointers, what means that even built-in types have to
  be allocated on the heap to be inserted into the list, which is time
  and space consuming.
* It requires space for an additional pointer to the data for each
  element inserted in the list.
* It is not type-safe: list elements are cast to void* before insertion
  and then should be cast back to a correct type by a client, which is
  error-prone.
* Lists that hold values of different types have the same type List*,
  what causes confusion in client code.

CList is a type-safe list that can hold elements by value instead of
holding them by a pointer to the heap. CList has none of the List`s
drawbacks and offers much cleaner API for the clients.
CList is not only much more efficient than List but also produces
compile-time errors if the client inserts an element of the wrong type
or confuses the CLists of different types.