Added journaling system
Added journaling to record every GA operation.
The journaling system allows extensibility by registering different
journals that will analyze information and display it in various ways.
DeathManager is replaced by the journaling system and is removed.

Improved the usage of the logging system, marked the start and the end
of each crucial function with logging macros.