Implemented GUI for Evolution
Added the interface to provide the statistics
of the evolution process in real-time.

GUI register callbacks in Journal component and thus
gets notified of all the steps of the GA.
GUI callbacks collect the data of interest and post it within the
message to the thread-safe Qt queue.
The Qt thread than grubs messages from the queue and plots the data.

GUI allows launching and stopping evolution processes. On launching, the
user can adjust all the GA parameters and select GA operators.
It worth noting that several evolutions can be run at once and
comfortably managed through GUI. Every evolution is run in the separate
thread in this case.