Refactored Species Links Selection
Selection based on links between species was leaving too many entities
alive, causing an enormous growth in the population size.
The population growth was caused by the selection limitations: it did
not kill entities in small species. Thus, with the growth of the number
of small species, the size of the population was growing too.

Species Links Selection now does not have these limitations and has been
refactored to reuse the same template as all selection algorithms that
are based on entities fitness.