Added Fitness-Norm chart, autoscaling
Added a chart, that shows the dependency of entity's fitness on its
norm. This chart is aimed to demonstrate how do entities get stuck in
the local maximums. This chart is called Fitness-Norm chart.

Implemented autoscaling for charts that are completely redrawn on each
iteration: Species, Fitness landscape and Fitness-Norm.
This is done to get a more detailed view of the charts` data and to
display data even if the data does not fit in the specified display ranges.