Fixed chromosome size bug
In EvolutionWorker, parameters.chromosome_size was used
to determine the size of the chromosomes in the world.
But for some objectives, the real size is smaller than
the size specified in the paraemetrs, because these
objectives restrict the maximum number of their arguments.
For these objectives, the code acessed the data beyond
the chromosome borders when counted entities` norms.