Implemented GraphWidget
Added GraphWidget that is responsible for building graphs of the objective functions.
At the moment, the graph is a fitness-norm graph for functions of all dimensions.
The graph is built by iterating over every point that should be drawn,
what is really slow and ineffective for dimensions higher than 6.
The data for the plot is generated in a separate thread of execution.