Fixed a bug in GAO_NonUniformMutation
GAO_NonUniformMutation function contained an error: it did not update
entity`s fitness value after mutating entity`s chromosome. The bug is
fixed and a debug check for chromosome and fitness consistency is added
to EvolutionWorker.

Removed copying of GAOperators. The GAOperators object is loaded by the
plugins system and stays available throughout the whole run of the
program. Thus, there is no need to copy it. Besides, holding this object
by const pointer avoids some errors with multithreading.

Refactored the code to collect the logic of determining the chromosome
size in one place.