Added zooming to CustomChartView
To let a user better explore graphs
a zooming feature has been added to all the charts.
Zooming is implemented through QChartView`s RectangleRubberBand,
which zooms in the selected area and zooms out on the right mouse click.
The zoom is removed on the mouse double-click.
The scrolling was not implemented, QChart::scroll() method
has been tried but it was producing ridiculous results.