Added Best-Worst selection
Added Best-Worst selection operator that leaves
both the best and the worst entities during selection.
The idea behind the operator is to leave entities
that strive to the direction opposite to the global maximum
so they could overcome the ravines produced by local minimums.
The crossover between the best and the worst entities
should give the entities outside the ravines
on the way to the global maximum.

The Best-Worst-Penalty operators have been added that start
with Herrera + Clustering operators and than switch to
Herrera + Clustering + Best-Worst selection + Cluster size penalty.