Added Scattering Selection
Scattering selection operator makes each species
to move away from the nearest species.

This is done by altering the entities` fitness.
The entity fitness is influenced by the direction
from the species` centroid to the entity:
the more this direction coincides with the direction
opposite to the direction from the current centroid
to the nearest species` centroid,
the more entity`s fitness is increased.
Thus, entities that are farther from the nearest species
become more adapted and are left alive during selection.
Besides the direction, the distance to the nearest species
and its size also influences the change in the entity`s fitness.